refactor dashboard into components
